+++ modules/core/src/main/java/net/ashald/envfile/EnvVarsProvider.java
@@ -2,6 +2,7 @@ package net.ashald.envfile;
 
 import org.jetbrains.annotations.NotNull;
 
+import java.io.File;
 import java.io.IOException;
 import java.util.Map;
 
@@ -11,4 +12,6 @@ public interface EnvVarsProvider {
     @NotNull Map<String, String> process(@NotNull Map<String, String> runConfigEnv, String path, @NotNull Map<String, String> aggregatedEnv) throws EnvFileErrorException, IOException;
 
     boolean isEditable();
+
+    boolean isFileLocationValid(File file);
 }