+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning].
 
-## Unreleased
+## 3.1.1 - 2019-04-24
 
 ### Fixed